JACQ-491 Introduce simple edit for living plants
The simple edit of living plants also use the same
controller like the normal edit of living plants. There was
no different logic to develop so I used the same controller. In case
a different logic has to be implementend first a reactoring should be
done so an own controller for simpel edit can be introduced.

Signed-off-by: Hans Sowa <hans.sowa@ogs.at>